update plot script
